% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{normalize}
\alias{normalize}
\alias{row_normalize}
\title{Normalize a vector to sum to one.}
\usage{
normalize(x)

row_normalize(X)
}
\arguments{
\item{x}{vector of input data.}

\item{X}{matrix of input data.}
}
\value{
the input divided by its sum. 
For the row-wise version, each row is divided by its sum.
}
\description{
Divide a vector by its sum, resulting in a vector with
sum equal to one.
}
\note{
This function will return \code{NA} when any elements of the input
are \code{NA}. May return \code{Inf} if the elements sum
to zero.
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
