% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohenery.r
\name{ohenery-NEWS}
\alias{ohenery-NEWS}
\title{News for package 'ohenery':}
\description{
News for package \sQuote{ohenery}

\newcommand{\CRANpkg}{\href{https://cran.r-project.org/package=#1}{\pkg{#1}}}
\newcommand{\ohenery}{\CRANpkg{ohenery}}
}
\section{\ohenery{} Initial Version 0.1.0 (2018-10-01) }{

\itemize{
\item first CRAN release.
}
}

