% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohenery.r
\name{ohenery-NEWS}
\alias{ohenery-NEWS}
\title{News for package 'ohenery':}
\description{
News for package \sQuote{ohenery}

\newcommand{\CRANpkg}{\href{https://cran.r-project.org/package=#1}{\pkg{#1}}}
\newcommand{\ohenery}{\CRANpkg{ohenery}}
}
\section{\ohenery{} Version 0.1.4 (2025-??-??) }{

\itemize{
\item Adding regularization. #9
\item Adding hensmfit experts function. #10
\item Adding broom tidy and glance methods. #11
}
}

\section{\ohenery{} Version 0.1.3 (2025-04-19) }{

\itemize{
\item Fix issue with factor levels in predict. #8
}
}

\section{\ohenery{} Version 0.1.2 (2024-10-25) }{

\itemize{
\item Allow warm start in harsm and hensm.
}
}

\section{\ohenery{} Version 0.1.1 (2018-10-14) }{

\itemize{
\item Change default in harsm and hensm to use unnormalized weights,
correcting inference when not all finishes are observed.
}
}

\section{\ohenery{} Initial Version 0.1.0 (2018-10-01) }{

\itemize{
\item first CRAN release.
}
}

